% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{example_BAM}
\alias{example_BAM}
\title{Returns the path of a test BAM file}
\usage{
example_BAM()
}
\arguments{
\item{pkg}{The name of the example package (default "ompBAMExample")}
}
\value{
None.
}
\description{
Installs the ompBAMExample package included with the ompBAM package
}
\details{
This function installs the ompBAMExample package located inside
the 'examples' subfolder of ompBAM.
}
\examples{
# The directory containing the source code is given by the path here

print(system.file(file.path('examples', "ompBAMExample"), 
    package = 'ompBAM'))
# Install the ompBAMExample package

install_ompBAM_example()
}
