% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ompBAM.R
\name{install_ompBAM_example}
\alias{install_ompBAM_example}
\title{ompBAM Example}
\usage{
install_ompBAM_example()
}
\value{
None.
}
\description{
Installs the ompBAMExample package included with the ompBAM package
}
\details{
This function installs the ompBAMExample package located inside
the 'examples' subfolder of ompBAM. It uses devtools::load_all() to simulate
package creation. After running this function to compile ompBAMExample(), 
users can run the example functions contained therein.
}
\examples{
# The directory containing the source code is given by the path here

print(system.file(file.path('examples', "ompBAMExample"), 
    package = 'ompBAM'))
# Install the ompBAMExample package

install_ompBAM_example()
}
