% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{install_ParaBAM_example}
\alias{install_ParaBAM_example}
\title{ParaBAM Example}
\usage{
install_ParaBAM_example(pkg = "ParaBAMExample")
}
\arguments{
\item{pkg}{The name of the example package (default "ParaBAMExample")}
}
\value{
None.
}
\description{
Installs the ParaBAMExample package included with the ParaBAM package
}
\details{
This function installs the ParaBAMExample package located inside
the 'examples' subfolder of ParaBAM.
}
\examples{
# The directory containing the source code is given by the path here

print(system.file(file.path('examples', "ParaBAMExample"), 
    package = 'ParaBAM'))
# Install the ParaBAMExample package

install_ParaBAM_example()
}
