% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{NxtIRF-package}
\alias{NxtIRF-package}
\title{ParaBAM: C++ header library for parallel sequential reading
of BAM files}
\description{
ParaBAM provides a set of tools to facilitate developers to design
Rcpp / C++ based packages to read BAM files in a multi-threaded,
sequential approach. Currently, a multi-threaded solution is not
available to developers using Rhtslib.
}
\details{

}
\author{
Alex Chit Hei Wong
}
\keyword{package}
