% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{NxtIRF-package}
\alias{NxtIRF-package}
\title{ompBAM: C++ header library for parallel sequential reading
of BAM files}
\description{
ompBAM provides a set of tools to facilitate developers to design
Rcpp / C++ based packages to read BAM files in a multi-threaded,
sequential approach. Currently, a multi-threaded solution is not
available to developers using Rhtslib.
}
\details{
For a quick start guide and a full list of functionality, refer to
the vignette
\code{vignette("ompBAM-API-Docs", package = "ompBAM")}
}
\author{
Alex Chit Hei Wong
}
\keyword{package}
